+++ src/utils.h
@@ -231,7 +231,7 @@ extern int si_existe_archivo(char *nombre);
 
 extern long long int get_file_size(char *nombre);
 
-extern long int get_size_human_friendly(long long int tamanyo,char *sufijo);
+extern long long int get_size_human_friendly(long long int tamanyo,char *sufijo);
 
 extern int lee_archivo(char *nombre,char *buffer,int max_longitud);
 
@@ -717,7 +717,7 @@ extern int si_existe_editionnamegame(char *nombre_final);
 extern int util_extract_mdv(char *mdvname, char *dest_dir);
 extern int util_extract_hdf(char *hdfname, char *dest_dir);
 
-extern void util_save_file(z80_byte *origin, long int tamanyo_origen, char *destination_file);
+extern void util_save_file(z80_byte *origin, long long int tamanyo_origen, char *destination_file);
 extern void util_load_file_bytes(z80_byte *taperead,char *filename,int total_leer);
 
 extern void util_string_replace_char(char *s,char orig,char dest);
